requirements.txt
streamlit
requests
vertexai
google-cloud-documentai 
google-auth